Move the attachment url hook to the sanitization helper
Pure relocation. attachmentUriFilterHook, its constants and its comments
now live in helpers/sanitization_helper, next to the other sanitization
entry points; config/dom_purify imports them. Registration stays with the
other hooks, and so does ADD_URI_SAFE_ATTR, which is config.

Nests the unit test under test/javascript/unit/sanitization/, using the
src alias the other nested tests use.
